+++ src/main/java/it/gov/pagopa/pu/organization/model/taxonomy/TaxonomyCodeDTO.java
@@ -1,7 +1,8 @@
-package it.gov.pagopa.pu.organization.model;
+package it.gov.pagopa.pu.organization.model.taxonomy;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.IdClass;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,6 +11,7 @@ import lombok.Data;
 @AllArgsConstructor
 @Entity
 @Table(name = "taxonomy")
+@IdClass(TaxonomyCodeDTO.class)
 public class TaxonomyCodeDTO {
   @Id
   private String organizationType;
