+++ src/main/java/it/gov/pagopa/pu/organization/model/taxonomy/TaxonomyCollectionReasonDTO.java
@@ -1,7 +1,8 @@
-package it.gov.pagopa.pu.organization.model;
+package it.gov.pagopa.pu.organization.model.taxonomy;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.IdClass;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,14 +11,20 @@ import lombok.Data;
 @AllArgsConstructor
 @Entity
 @Table(name = "taxonomy")
-public class TaxonomyServiceTypeCodeDTO {
+@IdClass(TaxonomyCollectionReasonDTO.class)
+public class TaxonomyCollectionReasonDTO {
+
   @Id
   private String organizationType;
   private String organizationTypeDescription;
+  @Id
   private String macroAreaCode;
   private String macroAreaName;
   private String macroAreaDescription;
+  @Id
   private String serviceTypeCode;
   private String serviceType;
   private String serviceTypeDescription;
+  @Id
+  private String collectionReason;
 }
