+++ src/main/java/it/gov/pagopa/pu/organization/model/taxonomy/TaxonomyMacroAreaCodeDTO.java
@@ -2,6 +2,7 @@ package it.gov.pagopa.pu.organization.model.taxonomy;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.IdClass;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,16 +11,15 @@ import lombok.Data;
 @AllArgsConstructor
 @Entity
 @Table(name = "taxonomy")
-public class TaxonomyCollectionReasonDTO {
+@IdClass(TaxonomyMacroAreaCodeDTO.class)
+public class TaxonomyMacroAreaCodeDTO {
 
   @Id
   private String organizationType;
   private String organizationTypeDescription;
+  @Id
   private String macroAreaCode;
   private String macroAreaName;
   private String macroAreaDescription;
-  private String serviceTypeCode;
-  private String serviceType;
-  private String serviceTypeDescription;
-  private String collectionReason;
+
 }
