+++ src/main/java/it/gov/pagopa/pu/organization/model/taxonomy/TaxonomyServiceTypeCodeDTO.java
@@ -1,7 +1,8 @@
-package it.gov.pagopa.pu.organization.model;
+package it.gov.pagopa.pu.organization.model.taxonomy;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.IdClass;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,13 +11,17 @@ import lombok.Data;
 @AllArgsConstructor
 @Entity
 @Table(name = "taxonomy")
-public class TaxonomyMacroAreaCodeDTO {
-
+@IdClass(TaxonomyServiceTypeCodeDTO.class)
+public class TaxonomyServiceTypeCodeDTO {
   @Id
   private String organizationType;
   private String organizationTypeDescription;
+  @Id
   private String macroAreaCode;
   private String macroAreaName;
   private String macroAreaDescription;
-
+  @Id
+  private String serviceTypeCode;
+  private String serviceType;
+  private String serviceTypeDescription;
 }
